test(dirs): move example app and test db to a separate directory
This is to make a clean cut between tests and test utils.
Mocha is configured to run tests from the /test/* directory.
We don't want Mocha to run scripts which are not tests.
